@JayCesar | feat: adds LoggerModelTest and rename Sender Class to SenderService
